+++ client/src/lib/nav-catagories.js
@@ -1,4 +1,3 @@
-const { config } = require("chai");
 const {
   resourceGroupTf,
   kmsTf,